requirements.txt
streamlit==1.39.0
groq==0.11.0
python-dotenv==1.0.1
pandas==2.2.3
